refactor: move install-agent subcommand
yargs has nice support for encapsulating a command in its own module.
Use it for the install-agent subcommand.
