+++ packages/cli/src/cmds/agentInstaller/buildToolInstallerBase.js
