+++ packages/cli/src/cmds/agentInstaller/commandStruct.js
