+++ packages/cli/src/cmds/agentInstaller/installAgentStep.js
